Add PageBy Fix OrderBy Complete

+++ orm/src/main/java/com/resqlity/orm/consts/Pagination.java
@@ -1,5 +1,5 @@
 package com.resqlity.orm.consts;
 
 public class Pagination {
-    public static final int PageSize = 10;
+    public static final long PageSize = 10;
 }
